+++ apps/web/src/app/modules/register-form/register-form.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 
-import { SharedModule } from "../shared.module";
+import { SharedModule } from "../../shared/shared.module";
 
 import { RegisterFormComponent } from "./register-form.component";
