+++ apps/web/src/app/modules/vault-filter/vault-filter.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 
-import { SharedModule } from "../shared.module";
+import { SharedModule } from "../../shared/shared.module";
 
 import { CollectionFilterComponent } from "./components/collection-filter.component";
 import { FolderFilterComponent } from "./components/folder-filter.component";
