+++ apps/web/src/app/modules/vault/modules/organization-badge/organization-badge.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 
-import { SharedModule } from "../../../shared.module";
+import { SharedModule } from "../../../../shared/shared.module";
 
 import { OrganizationNameBadgeComponent } from "./organization-name-badge.component";
